@@ -1,36 +1,40 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <artifactId>demo-email</artifactId>
-  <version>1.0.0-SNAPSHOT</version>
-  <packaging>jar</packaging>
-
-    <name>demo-email</name>
-  <description>Demo project for Spring Boot</description>
-
   <parent>
     <groupId>com.xkcoding</groupId>
-    <artifactId>spring-boot-demo</artifactId>
+    <artifactId>demo-base</artifactId>
     <version>1.0.0-SNAPSHOT</version>
   </parent>
 
+  <modelVersion>4.0.0</modelVersion>
+
+  <artifactId>demo-base-email</artifactId>
+  <version>1.0.0-SNAPSHOT</version>
+  <packaging>jar</packaging>
+
+  <name>demo-base-email</name>
+  <description>Demo project for Spring Boot</description>
+
+
   <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-    <java.version>1.8</java.version>
-    <jasypt.version>2.1.1</jasypt.version>
+    <java.version>17</java.version>
+    <jasypt.version>3.0.4</jasypt.version>
   </properties>
 
   <dependencies>
+    <dependency>
+      <groupId>com.xkcoding</groupId>
+      <artifactId>common-tools</artifactId>
+    </dependency>
+
     <!-- Spring Boot 邮件依赖 -->
     <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter-mail</artifactId>
     </dependency>
 
-    <!--jasypt配置文件加解密-->
+    <!-- jasypt配置文件加解密 -->
     <dependency>
       <groupId>com.github.ulisesbocchio</groupId>
       <artifactId>jasypt-spring-boot-starter</artifactId>
@@ -43,11 +47,6 @@
       <scope>test</scope>
     </dependency>
 
-    <dependency>
-      <groupId>cn.hutool</groupId>
-      <artifactId>hutool-all</artifactId>
-    </dependency>
-
     <!-- Spring Boot 模板依赖 -->
     <dependency>
       <groupId>org.springframework.boot</groupId>
@@ -56,7 +55,7 @@
   </dependencies>
 
   <build>
-      <finalName>demo-email</finalName>
+    <finalName>demo-base-email</finalName>
     <plugins>
       <plugin>
         <groupId>org.springframework.boot</groupId>
